Handling and sending space update
- Replaced map with space
- Updating protobuf spec
- Event dispatcher handles events and time ticks properly
- Added capacity to event queue so it doesn't block

+++ types/vector.go
@@ -2,12 +2,12 @@ package types
 
 // Vector struct
 type Vector struct {
-	X int32
-	Y int32
+	X float32
+	Y float32
 }
 
 // NewVector initialization
-func NewVector(x, y int32) *Vector {
+func NewVector(x, y float32) *Vector {
 	return &Vector{
 		X: x,
 		Y: y,
